+++ src/core/schemas/luma-asset.ts
@@ -2,9 +2,11 @@ import * as zod from "zod";
 
 export const LumaAssetUrlSchema = zod.string().url("Invalid luma url format.");
 
-export const LumaAssetSchema = zod.object({
-	type: zod.literal("luma"),
-	src: LumaAssetUrlSchema
-});
+export const LumaAssetSchema = zod
+	.object({
+		type: zod.literal("luma"),
+		src: LumaAssetUrlSchema
+	})
+	.strict();
 
 export type LumaAsset = zod.infer<typeof LumaAssetSchema>;