+++ src/core/schemas/shape-asset.ts
@@ -2,29 +2,39 @@ import * as zod from "zod";
 
 export const ShapeAssetColorSchema = zod.string().regex(/^#([A-Fa-f0-9]{8}|[A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})|transparent$/, "Invalid color format.");
 
-export const ShapeAssetRectangleSchema = zod.object({
-	width: zod.number().positive(),
-	height: zod.number().positive()
-});
+export const ShapeAssetRectangleSchema = zod
+	.object({
+		width: zod.number().positive(),
+		height: zod.number().positive()
+	})
+	.strict();
 
-export const ShapeAssetCircleSchema = zod.object({
-	radius: zod.number().positive()
-});
+export const ShapeAssetCircleSchema = zod
+	.object({
+		radius: zod.number().positive()
+	})
+	.strict();
 
-export const ShapeAssetLineSchema = zod.object({
-	length: zod.number().positive(),
-	thickness: zod.number().positive()
-});
+export const ShapeAssetLineSchema = zod
+	.object({
+		length: zod.number().positive(),
+		thickness: zod.number().positive()
+	})
+	.strict();
 
-export const ShapeAssetFillSchema = zod.object({
-	color: ShapeAssetColorSchema,
-	opacity: zod.number().min(0).max(1)
-});
+export const ShapeAssetFillSchema = zod
+	.object({
+		color: ShapeAssetColorSchema,
+		opacity: zod.number().min(0).max(1)
+	})
+	.strict();
 
-export const ShapeAssetStrokeSchema = zod.object({
-	color: ShapeAssetColorSchema,
-	width: zod.number().positive()
-});
+export const ShapeAssetStrokeSchema = zod
+	.object({
+		color: ShapeAssetColorSchema,
+		width: zod.number().positive()
+	})
+	.strict();
 
 export const ShapeAssetSchema = zod
 	.object({
@@ -38,6 +48,7 @@ export const ShapeAssetSchema = zod
 		circle: ShapeAssetCircleSchema.optional(),
 		line: ShapeAssetLineSchema.optional()
 	})
+	.strict()
 	.refine(schema => {
 		if (schema.shape === "rectangle") {
 			return ShapeAssetRectangleSchema.safeParse(schema.rectangle);