+++ src/core/schemas/track.ts
@@ -2,6 +2,8 @@ import * as zod from "zod";
 
 import { ClipSchema } from "./clip";
 
-export const TrackSchema = zod.object({
-	clips: ClipSchema.array()
-});
+export const TrackSchema = zod
+	.object({
+		clips: ClipSchema.array()
+	})
+	.strict();